+++ core-ng/src/main/java/core/framework/internal/stat/Stats.java
@@ -4,6 +4,7 @@ import core.framework.util.ASCII;
 import core.framework.util.Maps;
 
 import java.text.NumberFormat;
+import java.util.HashMap;
 import java.util.Map;
 
 /**
@@ -13,6 +14,7 @@ public class Stats {
     public final Map<String, Double> stats = Maps.newLinkedHashMap(); // to keep order in es
     public String errorCode;
     public String errorMessage;
+    public Map<String, String> info;
 
     public String result() {
         if (errorCode == null) return "OK";
@@ -23,11 +25,20 @@ public class Stats {
         stats.put(key, value);
     }
 
-    public void checkHighUsage(double usage, double threshold, String name) {
+    public boolean checkHighUsage(double usage, double threshold, String name) {
         if (usage >= threshold) {
             NumberFormat format = NumberFormat.getPercentInstance();
             errorCode = "HIGH_" + ASCII.toUpperCase(name) + "_USAGE";
             errorMessage = name + " usage is too high, usage=" + format.format(usage);
+            return true;
         }
+        return false;
+    }
+
+    public void info(String key, String value) {
+        if (info == null) {
+            info = new HashMap<>();
+        }
+        info.put(key, value);
     }
 }