+++ core-ng/src/main/java/core/framework/internal/web/management/DiagnosticController.java
@@ -1,16 +1,9 @@
 package core.framework.internal.web.management;
 
+import core.framework.internal.stat.Diagnostic;
 import core.framework.internal.web.http.IPv4AccessControl;
 import core.framework.web.Request;
 import core.framework.web.Response;
-
-import javax.management.InstanceNotFoundException;
-import javax.management.MBeanException;
-import javax.management.MBeanServer;
-import javax.management.MalformedObjectNameException;
-import javax.management.ObjectName;
-import javax.management.ReflectionException;
-import java.lang.management.ManagementFactory;
 
 /**
  * @author neo
@@ -22,28 +15,16 @@ public class DiagnosticController {
     // enabling NMT will result in a 5-10 percent JVM performance drop
     public Response vm(Request request) {
         accessControl.validate(request.clientIP());
-        return Response.text(invoke("vmInfo"));
+        return Response.text(Diagnostic.vm());
     }
 
     public Response thread(Request request) {
         accessControl.validate(request.clientIP());
-        return Response.text(invoke("threadPrint", "-e"));
+        return Response.text(Diagnostic.thread());
     }
 
     public Response heap(Request request) {
         accessControl.validate(request.clientIP());
-        return Response.text(invoke("gcClassHistogram"));
-    }
-
-    // use "jcmd pid help" to list all operations,
-    // refer to com.sun.management.internal.DiagnosticCommandImpl.getMBeanInfo, all command names are transformed
-    private String invoke(String operation, String... params) {
-        try {
-            MBeanServer server = ManagementFactory.getPlatformMBeanServer();
-            var name = new ObjectName("com.sun.management", "type", "DiagnosticCommand");
-            return (String) server.invoke(name, operation, new Object[]{params}, new String[]{String[].class.getName()});
-        } catch (MalformedObjectNameException | InstanceNotFoundException | MBeanException | ReflectionException e) {
-            throw new Error(e);
-        }
+        return Response.text(Diagnostic.heap());
     }
 }